+++ src/main/java/umc/catchy/domain/course/api/CourseController.java
@@ -9,7 +9,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import umc.catchy.domain.course.domain.CourseType;
 import umc.catchy.domain.course.dto.request.CourseCreateRequest;
 import umc.catchy.domain.course.dto.request.CourseUpdateRequest;
 import umc.catchy.domain.course.dto.response.CourseDetailResponse;
@@ -23,8 +22,6 @@ import umc.catchy.domain.course.service.CourseService;
 import umc.catchy.domain.courseReview.dto.request.PostCourseReviewRequest;
 import umc.catchy.domain.courseReview.dto.response.PostCourseReviewResponse;
 import umc.catchy.domain.courseReview.service.CourseReviewService;
-import umc.catchy.domain.mapping.memberCourse.dto.response.CourseBookmarkResponse;
-import umc.catchy.domain.mapping.memberCourse.dto.response.MemberCourseSliceResponse;
 import umc.catchy.domain.mapping.memberCourse.service.MemberCourseService;
 import umc.catchy.domain.mapping.placeVisit.dto.response.PlaceVisitedResponse;
 import umc.catchy.domain.mapping.placeVisit.service.PlaceVisitService;
@@ -50,7 +47,6 @@ public class CourseController {
     private final AICourseGenerationService aiCourseGenerationService;
     private final CourseRecommendationService courseRecommendationService;
     private final CourseReviewService courseReviewService;
-    private final MemberCourseService memberCourseService;
     private final PlaceService placeService;
     private final PlaceVisitService placeVisitService;
 
@@ -61,19 +57,6 @@ public class CourseController {
             @PathVariable Long courseId
     ){
         CourseDetailResponse response = courseService.getCourseDetails(courseId);
-        return ResponseEntity.ok(BaseResponse.onSuccess(SuccessStatus._OK, response));
-    }
-
-    @Operation(summary = "내 코스 조회 API", description = "코스 탭에서 DIY/AI, 지역별로 사용자의 코스를 최신순으로 조회")
-    @GetMapping("/search")
-    public ResponseEntity<BaseResponse<MemberCourseSliceResponse>> getMemberCourses(
-            @Parameter(description = "AI/DIY 선택", required = true)
-            @RequestParam(value = "type") CourseType type,
-            @RequestParam(value = "upperLocation", defaultValue = "all") String upperLocation,
-            @RequestParam(value = "lowerLocation", defaultValue = "all") String lowerLocation,
-            @RequestParam(required = false) Long lastId
-    ) {
-        MemberCourseSliceResponse response = courseService.getMemberCourses(type, upperLocation, lowerLocation, lastId);
         return ResponseEntity.ok(BaseResponse.onSuccess(SuccessStatus._OK, response));
     }
 
@@ -117,13 +100,6 @@ public class CourseController {
             request.setImages(Collections.emptyList());
         }
         PostCourseReviewResponse.newCourseReviewResponseDTO response = courseReviewService.postNewCourseReview(courseId, request);
-        return ResponseEntity.ok(BaseResponse.onSuccess(SuccessStatus._OK, response));
-    }
-
-    @Operation(summary = "코스 북마크 API", description = "사용자가 해당 코스를 북마크합니다.")
-    @PatchMapping("/{courseId}/bookmark")
-    public ResponseEntity<BaseResponse<CourseBookmarkResponse>> toggleBookmark(@PathVariable Long courseId) {
-        CourseBookmarkResponse response = memberCourseService.toggleBookmark(courseId);
         return ResponseEntity.ok(BaseResponse.onSuccess(SuccessStatus._OK, response));
     }
